+++ BackEnd/src/main/model/statistics/Point.java
@@ -1,10 +1,10 @@
 package main.model.statistics;
 
 public class Point {
-    public double x;
-    public double y;
+    public float x;
+    public float y;
 
-    public Point(double x, double y) {
+    public Point(float x, float y) {
         this.x = x;
         this.y = y;
     }
